refactor: split cli unix c/s codes
