+++ crates/openconnect-cli/src/client/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod config;
+pub(crate) mod state;
