+++ crates/openconnect-cli/src/client/state.rs
